refactor: clean-up
